requirements.txt
asgiref==3.4.1
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
countries==0.2.0
Django==3.2.7
django-countries==7.2.1
django-debug-toolbar==3.2.2
idna==3.2
isort==5.9.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
platformdirs==2.3.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
